travis: Don't upgrade homebrew.
Just install python and leave the rest at the default versions.

The default osx image on travis doesn't have an up-to-date
set of homebrew packages, so the `brew upgrade` step must
compile many packages from source and takes tens of minutes
to complete. This makes for very slow test feedback and
occasionally failures when the job times out.