Convert test_atflexer.py to pytest.
This file was using the `unittest` framework from the standard
library, but we invoke it through pytest. Remove the class
inheriting from unittest.TestCase and promote the methods
to module-level functions, which pytest's introspection
will still find.

The lexer member variable is replaced with a pytest fixture.

This removes a level of indirection (and indentation) and
makes it possible to use more advanced pytest features
like parametrization.